// src/common/IpcEvents.enum.ts
export enum IpcEvents {
	CreateWaitingForConnectionSharingSession = 'create-waiting-for-connection-sharing-session',
	SetPendingConnectionDevice = 'set-pending-connection-device',
	UnmarkRoomIDAsTaken = 'unmark-room-id-as-taken',
	GetAppPath = 'get-app-path',
	ResetWaitingForConnectionSharingSession = 'reset-waiting-for-connection-sharing-session',
	SetDeviceConnectedStatus = 'set-device-connected-status',
	DisconnectPeerAndDestroySharingSessionBySessionID = 'disconnect-peer-and-destroy-sharing-session-by-session-id',
	GetDesktopCapturerSourceIdBySharingSessionId = 'get-desktop-capturer-source-id-by-sharing-session-id',
	GetConnectedDevices = 'get-connected-devices-list',
	DisconnectDeviceById = 'disconnect-device-by-id',
	ReleaseDeviceConnection = 'release-device-connection',
	DisconnectAllDevices = 'disconnect-all-devices',
	GetViewerConnectionAvailability = 'get-viewer-connection-availability',
	ViewerConnectionAvailabilityChanged = 'viewer-connection-availability-changed',
	AppLanguageChanged = 'app-language-changed',
	GetDesktopCapturerServiceSourcesMap = 'get-desktop-capturer-service-sources-map',
	GetDesktopCapturerServiceSourcesByIds = 'get-desktop-capturer-service-sources-by-ids',
	GetWaitingForConnectionSharingSessionSourceId = 'get-waiting-for-connection-sharing-session-source-id',
	StartSharingOnWaitingForConnectionSharingSession = 'start-sharing-on-waiting-for-connection-sharing-session',
	GetPendingConnectionDevice = 'get-pending-connection-device',
	GetIsLinuxWaylandSession = 'get-is-linux-wayland-session',
	RequestDesktopCapturerPortalSource = 'request-desktop-capturer-portal-source',
	GetDesktopSharingSourceIds = 'get-desktop-sharing-source-ids',
	SetDesktopCapturerSourceId = 'set-desktop-capturer-source-id',
	GetAppLanguage = 'get-app-language',
	GetIsFirstTimeAppStart = 'get-is-not-first-time-app-start',
	SetAppStartedOnce = 'set-app-started-once',
	DestroySharingSessionById = 'destroy-sharing-session-by-id',
	GetPort = 'get-port',
	OpenExternalLink = 'open-external-link',
	WriteTextToClipboard = 'write-text-to-clipboard',
	CheckScreenRecordingPermission = 'check-screen-recording-permission',
	RelaunchApp = 'relaunch-app',
}